requirements.txt
selenium>=4.0.0
webdriver-manager>=4.0.0
pandas>=1.3.0
python-slugify>=8.0.0
requests>=2.28.0
Pillow>=8.0.0
streamlit>=1.28.0
openai>=1.0.0
python-dotenv>=1.0.0
tiktoken>=0.5.0